+++ CHANGELOG.md
@@ -23,6 +23,9 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#2207 When eager loading ids the query must be duped to avoid multiple
+  iteration problems not getting the required fields.
+
 * \#2203 Map/Reduce now works properly in conjunction with `Model#with`.
 
         Band.